+++ Source/Modules/High/AppModuleLauncher/Launcher.cpp
@@ -11,7 +11,7 @@
 #include "Low/Window.h"
 #include "Medium/Imgui/UI.h"
 
-#include "AppModuleLauncher.h"
+#include "Launcher.h"
 
 
 // To make an app "high" module available for usage, add include here
